Fix CommonJS `default` export
